Refactor commands to use `Effect.fn` and contextual dependencies
Commands now use `Effect.fn` and yield dependencies from context instead
of receiving them as closure parameters.

Inline commands are extracted to `src/commands/`. The `registerCommand`
signature changes to accept `() => Effect` so commands can construct
their runtime at invocation.
