+++ extension/src/commands/restartLsp.ts
@@ -1,0 +1,7 @@
+import { Effect } from "effect";
+import { LanguageClient } from "../services/LanguageClient.ts";
+
+export const restartLsp = Effect.fn(function* () {
+  const client = yield* LanguageClient;
+  yield* client.restart();
+});
