+++ extension/src/commands/runStale.ts
@@ -1,0 +1,50 @@
+import { Effect, flow, Option } from "effect";
+import { MarimoNotebookDocument } from "../schemas.ts";
+import { VsCode } from "../services/VsCode.ts";
+import { showErrorAndPromptLogs } from "../utils/showErrorAndPromptLogs.ts";
+
+export const runStale = Effect.fn(
+  function* () {
+    const code = yield* VsCode;
+    const notebook = Option.filterMap(
+      yield* code.window.getActiveNotebookEditor(),
+      (editor) => MarimoNotebookDocument.tryFrom(editor.notebook),
+    );
+
+    if (Option.isNone(notebook)) {
+      yield* showErrorAndPromptLogs(
+        "Must have an open marimo notebook to run stale cells.",
+      );
+      return;
+    }
+
+    const staleCells = notebook.value.getCells().filter((cell) => cell.isStale);
+
+    if (staleCells.length === 0) {
+      yield* Effect.logInfo("No stale cells found");
+      yield* code.window.showInformationMessage("No stale cells to run");
+      return;
+    }
+
+    yield* Effect.logInfo("Running stale cells").pipe(
+      Effect.annotateLogs({
+        staleCount: staleCells.length,
+        notebook: notebook.value.uri.toString(),
+      }),
+    );
+
+    // Execute stale cells using VS Code's notebook execution command
+    yield* code.commands.executeCommand("notebook.cell.execute", {
+      ranges: staleCells.map((cell) => ({
+        start: cell.index,
+        end: cell.index + 1,
+      })),
+    });
+  },
+  flow(
+    Effect.tapErrorCause(Effect.logError),
+    Effect.catchAllCause(() =>
+      showErrorAndPromptLogs("Failed to run stale cells."),
+    ),
+  ),
+);
