+++ extension/src/services/__tests__/ControllerRegistry.test.ts
@@ -19,7 +19,7 @@ const TestLanguageClientMock = Layer.succeed(
   LanguageClient,
   LanguageClient.make({
     channel: { name: "marimo-lsp", show() {} },
-    restart: Effect.void,
+    restart: () => Effect.void,
     executeCommand() {
       return Effect.void;
     },
